+++ custom-dapps/template/entrypoint.sh
@@ -12,4 +12,4 @@
 # specific language governing permissions and limitations under the License.
 
 set -e
-python3 template.py
+rollup-init python3 template.py